refactor: move SFT entrypoint to main_sft.py
Extract sft_entrypoint() and main() from sft_trainer.py into the new
skyrl/train/main_sft.py module, keeping sft_trainer.py focused on the
SFTTrainer class. Removes unused imports (sys, validate_sft_cfg,
initialize_ray) from sft_trainer.py.
